requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==4.0
django-heroku==0.3.1
django-phonenumber-field==6.0.0
gunicorn==20.1.0
phonenumberslite==8.12.41
Pillow==9.0.0
psycopg2==2.9.2
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0
